Consolidate macro tests into single file
